added some error handles for controllers

+++ src/main/java/pl/databucket/repository/RoleRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import pl.databucket.entity.Role;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Short> {
     Role findByName(String name);
     boolean existsByName(String name);
 }
